+++ pueue/src/daemon/callbacks.rs
@@ -9,8 +9,7 @@ use pueue_lib::{
 };
 
 use crate::{
-    daemon::state_helper::LockedState,
-    internal_prelude::*,
+    daemon::internal_state::state::LockedState, internal_prelude::*,
     process_helper::compile_shell_command,
 };