+++ pueue/src/daemon/network/message_handler/enqueue.rs
@@ -3,22 +3,21 @@ use pueue_lib::{
     format::format_datetime,
     network::message::*,
     settings::Settings,
-    state::SharedState,
     success_msg,
     task::{Task, TaskStatus},
 };
 
-use crate::daemon::network::response_helper::*;
+use crate::daemon::{internal_state::SharedState, network::response_helper::*};
 
 /// Invoked when calling `pueue enqueue`.
 /// Enqueue specific stashed tasks.
 pub fn enqueue(settings: &Settings, state: &SharedState, message: EnqueueMessage) -> Response {
     let mut state = state.lock().unwrap();
     // Get the affected task ids, based on the task selection.
-    let selected_task_ids = match message.tasks {
+    let selected_tasks = match message.tasks {
         TaskSelection::TaskIds(ref task_ids) => state
-            .tasks
-            .iter()
+            .tasks_mut()
+            .iter_mut()
             .filter(|(task_id, task)| {
                 if !task_ids.contains(task_id) {
                     return false;
@@ -29,11 +28,10 @@ pub fn enqueue(settings: &Settings, state: &SharedState, message: EnqueueMessage
                     TaskStatus::Stashed { .. } | TaskStatus::Locked { .. }
                 )
             })
-            .map(|(task_id, _)| *task_id)
-            .collect::<Vec<usize>>(),
+            .collect::<Vec<(&usize, &mut Task)>>(),
         TaskSelection::Group(ref group) => state
-            .tasks
-            .iter()
+            .tasks_mut()
+            .iter_mut()
             .filter(|(_, task)| {
                 if task.group != *group {
                     return false;
@@ -44,25 +42,20 @@ pub fn enqueue(settings: &Settings, state: &SharedState, message: EnqueueMessage
                     TaskStatus::Stashed { .. } | TaskStatus::Locked { .. }
                 )
             })
-            .map(|(task_id, _)| *task_id)
-            .collect::<Vec<usize>>(),
+            .collect::<Vec<(&usize, &mut Task)>>(),
         TaskSelection::All => state
-            .tasks
-            .iter()
+            .tasks_mut()
+            .iter_mut()
             .filter(|(_, task)| {
                 matches!(
                     task.status,
                     TaskStatus::Stashed { .. } | TaskStatus::Locked { .. }
                 )
             })
-            .map(|(task_id, _)| *task_id)
-            .collect::<Vec<usize>>(),
+            .collect::<Vec<(&usize, &mut Task)>>(),
     };
 
-    for task_id in &selected_task_ids {
-        // We just checked that they're there and the state is locked. It's safe to unwrap.
-        let task = state.tasks.get_mut(task_id).expect("Task should be there.");
-
+    for (_, task) in selected_tasks {
         // Either specify the point of time the task should be enqueued or enqueue the task
         // immediately.
         if message.enqueue_at.is_some() {