+++ pueue/src/daemon/network/message_handler/env.rs
@@ -1,7 +1,7 @@
-use pueue_lib::{network::message::*, settings::Settings, state::SharedState};
+use pueue_lib::{network::message::*, settings::Settings};
 
 use crate::{
-    daemon::{network::message_handler::ok_or_failure_message, state_helper::save_state},
+    daemon::{internal_state::SharedState, network::message_handler::ok_or_failure_message},
     ok_or_save_state_failure,
 };
 
@@ -18,7 +18,7 @@ pub fn env(settings: &Settings, state: &SharedState, message: EnvMessage) -> Res
             key,
             value,
         } => {
-            let Some(task) = state.tasks.get_mut(&task_id) else {
+            let Some(task) = state.tasks_mut().get_mut(&task_id) else {
                 return create_failure_response(format!("No task with id {task_id}"));
             };
 
@@ -31,7 +31,7 @@ pub fn env(settings: &Settings, state: &SharedState, message: EnvMessage) -> Res
             create_success_response("Environment variable set.")
         }
         EnvMessage::Unset { task_id, key } => {
-            let Some(task) = state.tasks.get_mut(&task_id) else {
+            let Some(task) = state.tasks_mut().get_mut(&task_id) else {
                 return create_failure_response(format!("No task with id {task_id}"));
             };
 
@@ -50,7 +50,7 @@ pub fn env(settings: &Settings, state: &SharedState, message: EnvMessage) -> Res
 
     // Save the state if there were any changes.
     if let Response::Success(_) = message {
-        ok_or_save_state_failure!(save_state(&state, settings));
+        ok_or_save_state_failure!(state.save(settings));
     }
 
     message