+++ pueue/src/daemon/network/message_handler/restart.rs
@@ -1,15 +1,16 @@
-use std::sync::MutexGuard;
-
 use chrono::Local;
 use pueue_lib::{
     aliasing::insert_alias,
     network::message::*,
     settings::Settings,
-    state::{SharedState, State},
     task::{Task, TaskStatus},
 };
 
-use crate::daemon::{network::response_helper::task_action_response_helper, process_handler};
+use crate::daemon::{
+    internal_state::{state::LockedState, SharedState},
+    network::response_helper::task_action_response_helper,
+    process_handler,
+};
 
 /// This is a small wrapper around the actual in-place task `restart` functionality.
 ///
@@ -50,14 +51,9 @@ pub fn restart_multiple(
 ///
 /// The "not in-place" restart functionality is actually just a copy the finished task + create a
 /// new task, which is completely handled on the client-side.
-fn restart(
-    state: &mut MutexGuard<State>,
-    to_restart: TaskToRestart,
-    stashed: bool,
-    settings: &Settings,
-) {
+fn restart(state: &mut LockedState, to_restart: TaskToRestart, stashed: bool, settings: &Settings) {
     // Check if we actually know this task.
-    let Some(task) = state.tasks.get_mut(&to_restart.task_id) else {
+    let Some(task) = state.tasks_mut().get_mut(&to_restart.task_id) else {
         return;
     };